Cache_ChemInfo/requirements.txt
django
pymysql
djangorestframework
django-filter
django-cors-headers
pillow
python-dotenv
django-storages
boto3
